[Refactor]: server action 한 파일에 관리
